Move account sql model to infrastructure
