+++ tests/test_database.py
@@ -8,7 +8,7 @@ from sqlalchemy import event, inspect, text
 from sqlmodel import Session, SQLModel, create_engine, select
 
 from niveshpy.database import _iregexp, get_session, initialize
-from niveshpy.models.account import Account
+from niveshpy.infrastructure.sqlite.models import Account
 from niveshpy.models.price import Price
 from niveshpy.models.security import Security, SecurityCategory, SecurityType
